Fixed issues caught by Xcode 5.1 b4.

+++ Source/UI/SDSearchDisplayController.m
@@ -224,8 +224,8 @@ static NSString *kSDSearchUserDefaultsKey = @"kSDSearchUserDefaultsKey";
         // set the searchbar text here.
         if (indexPath.row < [searchHistory count])
 		{
-            self.searchBar.text = [searchHistory objectAtIndex:indexPath.row];
-            self.selectedSearchItem = [searchHistory objectAtIndex:indexPath.row];
+            self.searchBar.text = [searchHistory objectAtIndex:(NSUInteger)indexPath.row];
+            self.selectedSearchItem = [searchHistory objectAtIndex:(NSUInteger)indexPath.row];
         }
 		else
 		{
@@ -305,7 +305,7 @@ static NSString *kSDSearchUserDefaultsKey = @"kSDSearchUserDefaultsKey";
             cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:identifier];
         
         if (indexPath.row < [searchHistory count])
-			cell.textLabel.text = [searchHistory objectAtIndex:indexPath.row];
+			cell.textLabel.text = [searchHistory objectAtIndex:(NSUInteger)indexPath.row];
 		else if(showsClearRecentSearchResultsRow)
 			cell.textLabel.text = @"Clear All Recent Searches";
 		else
